+++ sweble-wikitext-components-parent/swc-parser-lazy/src/main/java/org/sweble/wikitext/parser/preprocessor/WikitextPreprocessorState.java
@@ -19,6 +19,7 @@ package org.sweble.wikitext.parser.preprocessor;
 
 import org.sweble.wikitext.parser.ParserConfig;
 import org.sweble.wikitext.parser.WtEntityMap;
+import org.sweble.wikitext.parser.nodes.WtNode;
 
 import de.fau.cs.osr.ptk.common.ParserState;
 
@@ -148,6 +149,11 @@ public class WikitextPreprocessorState
 		this.hasOnlyInclude = hasOnlyInclude;
 	}
 
+	public boolean isConvertIllegalCodePoints()
+	{
+		return this.config.isConvertIllegalCodePoints();
+	}
+
 	public boolean isParseForInclusion()
 	{
 		return parseForInclusion;
@@ -183,5 +189,10 @@ public class WikitextPreprocessorState
 	public boolean isRedirectKeyword(String keyword)
 	{
 		return config.isRedirectKeyword(keyword);
+	}
+
+	public WtNode getEntity(int id)
+	{
+		return entityMap.getEntity(id);
 	}
 }